Kind of Implemented quest completion
Half assed the quest completation, now the items you collect perma stay in ur hand, and if you get one or another they got assigned locations instead of using array to go to one slot if that is open. This will cause issues later on so it better be fixed,
Tested object explosion to pixels code and its pure garbo, will need other ways to implement this.